fix: [#6] pass encoded URL to image proxy
We need to pass the original URL encoded URL:

http://index.torrust-demo.com/api/v1/proxy/image/https%3A%2F%2Fraw.githubusercontent.com%2Ftorrust%2Ftorrust-index%2Fdevelop%2Fdocs%2Fmedia%2Ftorrust_logo.png

from the Nginx reverse proxy to the Index API image proxy.

Otherwise we get a 404. The proxy was passing an URL like this:

"/v1/proxy/image/https:/raw.githubusercontent.com/torrust/torrust-index/develop/docs/media/torrust_logo.png

which is not valid because it containts unscaped chars.

Not It's passing:

"/v1/proxy/image/https%3A%2F%2Fraw.githubusercontent.com%2Ftorrust%2Ftorrust-index%2Fdevelop%2Fdocs%2Fmedia%2Ftorrust_logo.png"

This solution is not passing the rest of the URL: query parameters. But
we do not need them in the image proxy.